fix: move free-space estimate use case out of plans.rs (fence: #979)
plans.rs is being split on another in-flight branch (#979); move
estimate_free_space and its tests into a new sibling module,
plan_free_space.rs, so this PR carries zero net diff on plans.rs
itself. No behavior change.
